Package core/commoncmd provides shared functionality for omcmd and oxcmd
Updated `omcmd` and `oxcmd` commands to utilize the new shared implementation.

This commit also fix the previously ignored --node option on om node log
command.
